Add breadcrumbs and remove redundant navigation
Refactored base template 'extended_base.html' to display breadcrumbs.
Any template extending the base template now expects breadcrumbs to be
provided in the form of a list of dictionaries containing the text and
the url of the breadcrumb. The breadcrumb block of the base template
can be overridden to display custom contents or no contents at all. The
values of the breadcrumb context are sometimes hardcoded in the
respective view, and, in case of galleries, are prepared dynamically
using helper functions in galleries app.

Galleries, user login/registration pages, appointment creation pages
now have breadcrumbs in their headers.

The styles of the breadcrumbs are in '_breadcrumbs.scss'.

This commit also removes unnecessary navigation menu from gallery pages,
appointment pages, and user login/register pages to reduce clutter and
streamline the user experience. The navigation menu remains available on
the homepage.