Refactor templates and remove redundant files
Refactored templates so that they can be correctly extended. Previously,
the base template was created from three separate templates: header,
main, and footer. This created problems with extendability of the main
template. Now the base template is bigger and includes header and
footer.

The change made many files redundant, and they have been deleted in this
commit.

'logo.html' is a leftover file so it was deleted together with other
redundant or unnecessary files.