Logged-in users can use 'Create Appointment' page
Authenticated users can now create appointments on the 'Create
Appointment' page.

'create_appointment' view now uses a new form for authenticated users
that only has 'message' field. This form is displayed on
'Create Appointment' page below the calendar. After the user selects the
date of appointment and writes the (optional) messages, and submits the
form, the appointment is validated and created.