Send confirmation link before registration of user
The app is now sending a confirmation link to the mailbox of the user
that is attempting to register on the website. This link does not
expire. Registration is completed is user is assigned 'is_active=True'
status after the user clicks on the link. If the user does not click
on the link but instead attempts to register again (say, in case
they miss the email), a new link is generated and sent to the indicated
mailbox. The first link does not expire. This is to be refactored down
the road to avoid possible email spamming or similar abuse.