Users can delete their account from 'Client Area'
When a User clicks on a link in the 'Client Area' they go to the 'delete
user' page where they can confirm the deletion, and after that the user
is logged off and deactivated.

If the customer later changes their mind, they can reregister. This case
is handled by the same view that handles registration of new users, and
that customer will receive an email to confirm 'registration' (which is
actually just reactivation of the user). Upon clicking the link in the
email, the user will be reactivated.

Measures have been put in place in the client registration view to
obfuscate the actual status of the user for anyone trying to get access
through the registration menu. It is unclear, for instance, if the user
does not exist, or if the user has been deactivated.