Fix CSS columns - IntersectionObserver bug
This compensates for the bug that causes some images to fail to
lazy-load. The bug is caused by difficulties in using CSS 'columns'
layout with InterSection observer. After the page is loaded in addition
to running the IntersectionObserver with lazy-loading on all images,
the script waits and then checks for any cases where lazy-loading failed
within the viewport, and forces lazy-loading of any such images.